Set epsilon according to the primal tolerance
When epsilon is too small, infeasible problems are treated as feasible
and optimized.  This exacts a toll in performance.  Now, both epsilon
and the primal tolerance are set to 1e-10, and the code runs a lot
faster.

Thanks to Carleton Coffrin (@ccoffrin) for his help diagnosing the
source of the false-positive solutions and for proposing primal
tolerance as a means for preventing them.